zipit.sh: Add coinstatsindex
A demo is at https://pn.bublina.eu.org/bitcoin-mini.zip

OpenBSD signify(1) signature of the extracted files' checksums
comming soon. The idea is to extract on-the-fly and although
coinstatsindex should be enough for verifying the blockchain
and chainstate is OK, using signify you will be able to verify
that anyone signed it. Hahaha.